requirements.txt
customtkinter==5.2.2
markdown-it-py==4.0.0
tkinterweb==4.25.1
ollama==0.6.1